kitten-scientists: add ReligionTab
Tabs are extensions of TabManager, customized to include any method
necessary to call on the tab. In the simple cases it will just be the
specific click that is necessary. Other tabs, such as Bonfire or Trade
tabs will most likely just return an exact button match, but religion tab
only needs a specific click action.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>