kitten-scientists: enable toggling seasons for trading
This helps ensure we can easily control which seasons we trade the various
races in. There doesn't really need to be more control otherwise.

To make it display right, I have to reduce the width reduction in the mid
options div.

Also this is *clearly* very hackish, and could use a better integration.
The current options system is way too reliant on the getToggle/getOptions
flow so it is very difficult to extend it.

At any rate, that can be refactored in the future. This patch is very
useful for letting user control what seasons are used for each race.

It depends a lot on the refactor before this, so hopefully that is
acceptable as well.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>