resources: separate crafts from stocks
Not all items we need to stock are actually craftable. In addition, it
really doesn't belong inside the crafting options, but in a separate
section.

This allows us to put a modest 1000 furs stock limit on by default, so
that we don't use up all our furs with parchment crafting. This is useful
in the case where we want to do our best to ensure that we maintain some
furs for happiness.

The stock list only contains the current stock for now.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>