crafting: implement stock setting via dialog prompts
Add stocks option setting next to resources  in the craft section (since
this is where it makes the most sense visually). The user can click stocks
or items, and it will alternate display of the particular settings.

The stocks list includes a new button "add stock" which the user will be
prompted for a resource name to add a stock. This way, each user may
customize the stock list easily, and is not cluttered with several stocks
they do not care about. There is a delete button incase the user wants to
remove the item from the list, though setting it to zero will also
essentially remove it.

We use the game.getDisplayValueExt function.. however it does not really
allow us to read in the reverse values.. for now I am not sure best way to
solve this as users probably will want ability to input "1K" and so
forth. However, currently I can't think of an easy way to reverse
implement the game function.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>