options: shorten names and add borders on toggles with further options
Move the toggle functions inline, so that they are near the code that
actually defines the display of the checkbox, etc.

This patch adds borders around the item boxes so that it is easier to
distinguish between the various options quickly. This is valuable because
otherwise it becomes harder for the human eye to be sure we clicked on the
correct "items" toggle.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>