general: fix on-click triggers to only change state when it changed
the .change() setting can be triggered via jQuery, and we intend to add
functionality in a following patch to do just that. However this will
currently cause a large dump of messages due to "changing" to a state that
we already are at. Fix all on-click functions to only print messages when
their actual underlying state was changed.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>